r40 == r17_beta20(2012-06-03) 231_r40_kene.zip (斧のみ) 前回と殆ど変わっていません。
† 難易度を全体的には少し簡単の方向へ修正。
† その他。

+++ src/PSPL/video/PSPL_blit_1.c
@@ -107,7 +107,7 @@ static void Blit1to2(SDL_BlitInfo *info)
 	}
 	#else
 	/* Memory align at 4-byte boundary, if necessary */
-	if ( (long)dst & 0x03 )
+	if ( (u32)dst & 0x03 )
 	{
 		/* Don't do anything if width is 0 */
 		if ( width == 0 )
