+++ src/jpeg/jmemmgr.c
@@ -137,7 +137,7 @@ typedef struct
 	jvirt_barray_ptr virt_barray_list;
 
 	/* This counts total space obtained from jpeg_get_small/large */
-	long total_space_allocated;
+	s32 total_space_allocated;
 
 	/* alloc_sarray and alloc_barray set this value for use by virtual
 	 * array routines.
@@ -210,14 +210,14 @@ print_mem_stats (j_common_ptr cinfo, int pool_id)
 		lhdr_ptr = lhdr_ptr->hdr.next)
 	{
 		error(0/*ERR_FATAL*/, "  Large chunk used %ld\n",
-			(long) lhdr_ptr->hdr.bytes_used);
+			(s32) lhdr_ptr->hdr.bytes_used);
 	}
 	for (shdr_ptr = mem->small_list[pool_id]; shdr_ptr != NULL;
 	   shdr_ptr = shdr_ptr->hdr.next)
 	{
 		error(0/*ERR_FATAL*/, "  Small chunk used %ld free %ld\n",
-			(long) shdr_ptr->hdr.bytes_used,
-			(long) shdr_ptr->hdr.bytes_left);
+			(s32) shdr_ptr->hdr.bytes_used,
+			(s32) shdr_ptr->hdr.bytes_left);
 	}
 }
 
@@ -411,14 +411,14 @@ static const size_t extra_pool_slop[JPOOL_NUMPOOLS] =
   JSAMPARRAY result;
   JSAMPROW workspace;
   JDIMENSION rowsperchunk, currow, i;
-  long ltemp;
+  s32 ltemp;
 
   /* Calculate max # of rows allowed in one allocation chunk */
   ltemp = (MAX_ALLOC_CHUNK-SIZEOF(large_pool_hdr)) /
-		  ((long) samplesperrow * SIZEOF(JSAMPLE));
+		  ((s32) samplesperrow * SIZEOF(JSAMPLE));
   if (ltemp <= 0)
 	ERREXIT(cinfo, JERR_WIDTH_OVERFLOW);
-  if (ltemp < (long) numrows)
+  if (ltemp < (s32) numrows)
 	rowsperchunk = (JDIMENSION) ltemp;
   else
 	rowsperchunk = numrows;
@@ -459,14 +459,14 @@ static const size_t extra_pool_slop[JPOOL_NUMPOOLS] =
   JBLOCKARRAY result;
   JBLOCKROW workspace;
   JDIMENSION rowsperchunk, currow, i;
-  long ltemp;
+  s32 ltemp;
 
   /* Calculate max # of rows allowed in one allocation chunk */
   ltemp = (MAX_ALLOC_CHUNK-SIZEOF(large_pool_hdr)) /
-		  ((long) blocksperrow * SIZEOF(JBLOCK));
+		  ((s32) blocksperrow * SIZEOF(JBLOCK));
   if (ltemp <= 0)
 	ERREXIT(cinfo, JERR_WIDTH_OVERFLOW);
-  if (ltemp < (long) numrows)
+  if (ltemp < (s32) numrows)
 	rowsperchunk = (JDIMENSION) ltemp;
   else
 	rowsperchunk = numrows;
@@ -593,8 +593,8 @@ realize_virt_arrays (j_common_ptr cinfo)
 /* Allocate the in-memory buffers for any unrealized virtual arrays */
 {
   my_mem_ptr mem = (my_mem_ptr) cinfo->mem;
-  long space_per_minheight, maximum_space, avail_mem;
-  long minheights, max_minheights;
+  s32 space_per_minheight, maximum_space, avail_mem;
+  s32 minheights, max_minheights;
   jvirt_sarray_ptr sptr;
   jvirt_barray_ptr bptr;
 
@@ -606,18 +606,18 @@ realize_virt_arrays (j_common_ptr cinfo)
   maximum_space = 0;
   for (sptr = mem->virt_sarray_list; sptr != NULL; sptr = sptr->next) {
 	if (sptr->mem_buffer == NULL) { /* if not realized yet */
-	  space_per_minheight += (long) sptr->maxaccess *
-							 (long) sptr->samplesperrow * SIZEOF(JSAMPLE);
-	  maximum_space += (long) sptr->rows_in_array *
-					   (long) sptr->samplesperrow * SIZEOF(JSAMPLE);
+	  space_per_minheight += (s32) sptr->maxaccess *
+							 (s32) sptr->samplesperrow * SIZEOF(JSAMPLE);
+	  maximum_space += (s32) sptr->rows_in_array *
+					   (s32) sptr->samplesperrow * SIZEOF(JSAMPLE);
 	}
   }
   for (bptr = mem->virt_barray_list; bptr != NULL; bptr = bptr->next) {
 	if (bptr->mem_buffer == NULL) { /* if not realized yet */
-	  space_per_minheight += (long) bptr->maxaccess *
-							 (long) bptr->blocksperrow * SIZEOF(JBLOCK);
-	  maximum_space += (long) bptr->rows_in_array *
-					   (long) bptr->blocksperrow * SIZEOF(JBLOCK);
+	  space_per_minheight += (s32) bptr->maxaccess *
+							 (s32) bptr->blocksperrow * SIZEOF(JBLOCK);
+	  maximum_space += (s32) bptr->rows_in_array *
+					   (s32) bptr->blocksperrow * SIZEOF(JBLOCK);
 	}
   }
 
@@ -647,7 +647,7 @@ realize_virt_arrays (j_common_ptr cinfo)
 
   for (sptr = mem->virt_sarray_list; sptr != NULL; sptr = sptr->next) {
 	if (sptr->mem_buffer == NULL) { /* if not realized yet */
-	  minheights = ((long) sptr->rows_in_array - 1L) / sptr->maxaccess + 1L;
+	  minheights = ((s32) sptr->rows_in_array - 1L) / sptr->maxaccess + 1L;
 	  if (minheights <= max_minheights) {
 		/* This buffer fits in memory */
 		sptr->rows_in_mem = sptr->rows_in_array;
@@ -655,9 +655,9 @@ realize_virt_arrays (j_common_ptr cinfo)
 		/* It doesn't fit in memory, create backing store. */
 		sptr->rows_in_mem = (JDIMENSION) (max_minheights * sptr->maxaccess);
 		jpeg_open_backing_store(cinfo, & sptr->b_s_info,
-								(long) sptr->rows_in_array *
-								(long) sptr->samplesperrow *
-								(long) SIZEOF(JSAMPLE));
+								(s32) sptr->rows_in_array *
+								(s32) sptr->samplesperrow *
+								(s32) SIZEOF(JSAMPLE));
 		sptr->b_s_open = TRUE;
 	  }
 	  sptr->mem_buffer = alloc_sarray(cinfo, JPOOL_IMAGE,
@@ -671,7 +671,7 @@ realize_virt_arrays (j_common_ptr cinfo)
 
   for (bptr = mem->virt_barray_list; bptr != NULL; bptr = bptr->next) {
 	if (bptr->mem_buffer == NULL) { /* if not realized yet */
-	  minheights = ((long) bptr->rows_in_array - 1L) / bptr->maxaccess + 1L;
+	  minheights = ((s32) bptr->rows_in_array - 1L) / bptr->maxaccess + 1L;
 	  if (minheights <= max_minheights) {
 		/* This buffer fits in memory */
 		bptr->rows_in_mem = bptr->rows_in_array;
@@ -679,9 +679,9 @@ realize_virt_arrays (j_common_ptr cinfo)
 		/* It doesn't fit in memory, create backing store. */
 		bptr->rows_in_mem = (JDIMENSION) (max_minheights * bptr->maxaccess);
 		jpeg_open_backing_store(cinfo, & bptr->b_s_info,
-								(long) bptr->rows_in_array *
-								(long) bptr->blocksperrow *
-								(long) SIZEOF(JBLOCK));
+								(s32) bptr->rows_in_array *
+								(s32) bptr->blocksperrow *
+								(s32) SIZEOF(JBLOCK));
 		bptr->b_s_open = TRUE;
 	  }
 	  bptr->mem_buffer = alloc_barray(cinfo, JPOOL_IMAGE,
@@ -699,19 +699,19 @@ realize_virt_arrays (j_common_ptr cinfo)
 do_sarray_io (j_common_ptr cinfo, jvirt_sarray_ptr ptr, boolean writing)
 /* Do backing store read or write of a virtual sample array */
 {
-  long bytesperrow, file_offset, byte_count, rows, thisrow, i;
+  s32 bytesperrow, file_offset, byte_count, rows, thisrow, i;
 
-  bytesperrow = (long) ptr->samplesperrow * SIZEOF(JSAMPLE);
+  bytesperrow = (s32) ptr->samplesperrow * SIZEOF(JSAMPLE);
   file_offset = ptr->cur_start_row * bytesperrow;
   /* Loop to read or write each allocation chunk in mem_buffer */
-  for (i = 0; i < (long) ptr->rows_in_mem; i += ptr->rowsperchunk) {
+  for (i = 0; i < (s32) ptr->rows_in_mem; i += ptr->rowsperchunk) {
 	/* One chunk, but check for short chunk at end of buffer */
-	rows = psp_min((long) ptr->rowsperchunk, (long) ptr->rows_in_mem - i);
+	rows = psp_min((s32) ptr->rowsperchunk, (s32) ptr->rows_in_mem - i);
 	/* Transfer no more than is currently defined */
-	thisrow = (long) ptr->cur_start_row + i;
-	rows = psp_min(rows, (long) ptr->first_undef_row - thisrow);
+	thisrow = (s32) ptr->cur_start_row + i;
+	rows = psp_min(rows, (s32) ptr->first_undef_row - thisrow);
 	/* Transfer no more than fits in file */
-	rows = psp_min(rows, (long) ptr->rows_in_array - thisrow);
+	rows = psp_min(rows, (s32) ptr->rows_in_array - thisrow);
 	if (rows <= 0)				/* this chunk might be past end of file! */
 	  break;
 	byte_count = rows * bytesperrow;
@@ -731,20 +731,20 @@ do_sarray_io (j_common_ptr cinfo, jvirt_sarray_ptr ptr, boolean writing)
 /*LOCAL*/static void do_barray_io (j_common_ptr cinfo, jvirt_barray_ptr ptr, boolean writing)
 /* Do backing store read or write of a virtual coefficient-block array */
 {
-	long bytesperrow, file_offset, byte_count, rows, thisrow, i;
+	s32 bytesperrow, file_offset, byte_count, rows, thisrow, i;
 
-	bytesperrow = (long) ptr->blocksperrow * SIZEOF(JBLOCK);
+	bytesperrow = (s32) ptr->blocksperrow * SIZEOF(JBLOCK);
 	file_offset = ptr->cur_start_row * bytesperrow;
 	/* Loop to read or write each allocation chunk in mem_buffer */
-	for (i = 0; i < (long) ptr->rows_in_mem; i += ptr->rowsperchunk)
+	for (i = 0; i < (s32) ptr->rows_in_mem; i += ptr->rowsperchunk)
 	{
 		/* One chunk, but check for short chunk at end of buffer */
-		rows = psp_min((long) ptr->rowsperchunk, (long) ptr->rows_in_mem - i);
+		rows = psp_min((s32) ptr->rowsperchunk, (s32) ptr->rows_in_mem - i);
 		/* Transfer no more than is currently defined */
-		thisrow = (long) ptr->cur_start_row + i;
-		rows = psp_min(rows, (long) ptr->first_undef_row - thisrow);
+		thisrow = (s32) ptr->cur_start_row + i;
+		rows = psp_min(rows, (s32) ptr->first_undef_row - thisrow);
 		/* Transfer no more than fits in file */
-		rows = psp_min(rows, (long) ptr->rows_in_array - thisrow);
+		rows = psp_min(rows, (s32) ptr->rows_in_array - thisrow);
 		if (rows <= 0)				/* this chunk might be past end of file! */
 		{	break;	}
 		byte_count = rows * bytesperrow;
@@ -807,8 +807,8 @@ do_sarray_io (j_common_ptr cinfo, jvirt_sarray_ptr ptr, boolean writing)
 		else
 		{
 			/* use long arithmetic here to avoid overflow & unsigned problems */
-			long ltemp;
-			ltemp = (long) end_row - (long) ptr->rows_in_mem;
+			s32 ltemp;
+			ltemp = (s32) end_row - (s32) ptr->rows_in_mem;
 			if (ltemp < 0)
 			{	ltemp = 0;	}			/* don't fall off front end of file */
 			ptr->cur_start_row = (JDIMENSION) ltemp;
@@ -898,9 +898,9 @@ do_sarray_io (j_common_ptr cinfo, jvirt_sarray_ptr ptr, boolean writing)
 	  ptr->cur_start_row = start_row;
 	} else {
 	  /* use long arithmetic here to avoid overflow & unsigned problems */
-	  long ltemp;
+	  s32 ltemp;
 
-	  ltemp = (long) end_row - (long) ptr->rows_in_mem;
+	  ltemp = (s32) end_row - (s32) ptr->rows_in_mem;
 	  if (ltemp < 0)
 		ltemp = 0;				/* don't fall off front end of file */
 	  ptr->cur_start_row = (JDIMENSION) ltemp;
@@ -1052,7 +1052,7 @@ self_destruct (j_common_ptr cinfo)
 jinit_memory_mgr (j_common_ptr cinfo)
 {
   my_mem_ptr mem;
-  long max_to_use;
+  s32 max_to_use;
   int pool;
   size_t test_mac;
 
@@ -1073,7 +1073,7 @@ jinit_memory_mgr (j_common_ptr cinfo)
    * But a "constant too large" warning means you need to fix MAX_ALLOC_CHUNK.
    */
   test_mac = (size_t) MAX_ALLOC_CHUNK;
-  if ((long) test_mac != MAX_ALLOC_CHUNK ||
+  if ((s32) test_mac != MAX_ALLOC_CHUNK ||
 	  (MAX_ALLOC_CHUNK % SIZEOF(ALIGN_TYPE)) != 0)
 	ERREXIT(cinfo, JERR_BAD_ALLOC_CHUNK);
